Cleanup plugins directory code
