+++ content/backstage/plugins/incident-io.md
@@ -8,12 +8,9 @@ attribution:
   href: https://incident.io/
 
 intro: |
-  <p>
-    As a prerequisite, you need to have configured the Incident.io <a href="https://github.com/incident-io/catalog-importer">catalog-importer</a> to sync with your Backstage catalog. You can visit your incident.io dashboard to create a custom field that is powered by the Backstage Component catalog type.
-  </p>
-  <p>
-    We recommend creating a multi-select field called something like "Affected services" or "Impacted components".
-  </p>
+  As a prerequisite, you need to have configured the Incident.io [catalog-importer](https://github.com/incident-io/catalog-importer) to sync with your Backstage catalog. You can visit your incident.io dashboard to create a custom field that is powered by the Backstage Component catalog type.
+
+  We recommend creating a multi-select field called something like "Affected services" or "Impacted components".
 
 seo:
   title: 'Backstage Incident.io Plugin | Roadie'
