+++ src/components/backstage/plugins/index.js
@@ -3,3 +3,9 @@ export { default as Logo } from './Logo';
 export { default as ListItem } from './ListItem';
 export { default as Header } from './Header';
 export { default as EditOnGitHubLink } from './EditOnGitHubLink';
+export { default as PluginCTA } from './PluginCTA';
+export { default as Intro } from './Intro';
+export { default as CoverImage } from './CoverImage';
+export { default as InstallationSteps } from './InstallationSteps';
+export { default as Notes } from './Notes';
+export { default as PlaceholderBody } from './PlaceholderBody';
